refactor to partial
